# core-lending-loan-origination-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-loan-origination
  description: Loan Origination Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/loan-origination
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/loan-applications/{applicationId}/exceptions/{exceptionId}:
    get:
      tags:
        - ApplicationException
      summary: Get an exception
      description: Fetch a specific exception record by ID.
      operationId: getException
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: exceptionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApplicationExceptionDTO'
    put:
      tags:
        - ApplicationException
      summary: Update an exception
      description: Updates an existing exception record.
      operationId: updateException
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: exceptionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationExceptionDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApplicationExceptionDTO'
    delete:
      tags:
        - ApplicationException
      summary: Delete an exception
      description: Removes an exception record from the application.
      operationId: deleteException
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: exceptionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/v1/loan-applications/{applicationId}/documents/{documentId}:
    get:
      tags:
        - ApplicationDocument
      summary: Get a document
      description: Fetch a specific document record by ID.
      operationId: getDocument
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApplicationDocumentDTO'
    put:
      tags:
        - ApplicationDocument
      summary: Update a document record
      description: Updates metadata for an existing document.
      operationId: updateDocument
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationDocumentDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApplicationDocumentDTO'
    delete:
      tags:
        - ApplicationDocument
      summary: Delete a document record
      description: Removes a document record from the application.
      operationId: deleteDocument
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/v1/loan-applications/{applicationId}/exceptions:
    get:
      tags:
        - ApplicationException
      summary: List exceptions
      description: Retrieves a paginated list of exceptions for a loan application.
      operationId: findAllExceptions
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pageNumber
          in: query
          description: The zero-based page number to retrieve.
          required: false
          schema:
            type: string
            description: The zero-based page number to retrieve.
            example: 0
            default: '0'
          example: 0
        - name: pageSize
          in: query
          description: The number of items per page.
          required: false
          schema:
            type: string
            description: The number of items per page.
            example: 10
            default: '10'
          example: 10
        - name: sortBy
          in: query
          description: The field to sort the results by.
          required: false
          schema:
            type: string
            description: The field to sort the results by.
            example: name
          example: name
        - name: sortDirection
          in: query
          description: The direction of sorting, either ASC or DESC.
          required: false
          schema:
            type: string
            description: The direction of sorting, either ASC or DESC.
            example: DESC
            default: DESC
          example: DESC
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseApplicationExceptionDTO'
    post:
      tags:
        - ApplicationException
      summary: Add an exception
      description: Adds a new exception record for the application.
      operationId: createException
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationExceptionDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApplicationExceptionDTO'
  /api/v1/loan-applications/{applicationId}/documents:
    get:
      tags:
        - ApplicationDocument
      summary: List documents
      description: Retrieves a paginated list of documents for a loan application.
      operationId: findAllDocuments
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pageNumber
          in: query
          description: The zero-based page number to retrieve.
          required: false
          schema:
            type: string
            description: The zero-based page number to retrieve.
            example: 0
            default: '0'
          example: 0
        - name: pageSize
          in: query
          description: The number of items per page.
          required: false
          schema:
            type: string
            description: The number of items per page.
            example: 10
            default: '10'
          example: 10
        - name: sortBy
          in: query
          description: The field to sort the results by.
          required: false
          schema:
            type: string
            description: The field to sort the results by.
            example: name
          example: name
        - name: sortDirection
          in: query
          description: The direction of sorting, either ASC or DESC.
          required: false
          schema:
            type: string
            description: The direction of sorting, either ASC or DESC.
            example: DESC
            default: DESC
          example: DESC
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseApplicationDocumentDTO'
    post:
      tags:
        - ApplicationDocument
      summary: Add a document record
      description: Adds a new document record for the application.
      operationId: createDocument
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationDocumentDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApplicationDocumentDTO'
...
components:
  schemas:
...
    ApplicationExceptionDTO:
      required:
        - loanApplicationId
      type: object
      properties:
        exceptionId:
          type: string
          format: uuid
          readOnly: true
        loanApplicationId:
          type: string
          format: uuid
        exceptionType:
          maxLength: 50
          minLength: 0
          type: string
        exceptionReason:
          maxLength: 1000
          minLength: 0
          type: string
        requestedBy:
          type: string
        requestedAt:
          type: string
          format: date-time
        approvalStatus:
          maxLength: 50
          minLength: 0
          type: string
        approvedBy:
          type: string
        approvedAt:
          type: string
          format: date-time
        rejectionReason:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ApplicationDocumentDTO:
      required:
        - documentId
        - documentTypeId
        - isMandatory
        - isReceived
        - loanApplicationId
      type: object
      properties:
        applicationDocumentId:
          type: string
          format: uuid
          readOnly: true
        loanApplicationId:
          type: string
          format: uuid
        documentId:
          type: string
          format: uuid
        documentTypeId:
          type: string
          format: uuid
        isMandatory:
          type: boolean
        isReceived:
          type: boolean
        receivedAt:
          type: string
          format: date-time
        documentName:
          maxLength: 200
          minLength: 0
          type: string
        fileSizeBytes:
          type: integer
          format: int64
        mimeType:
          maxLength: 100
          minLength: 0
          type: string
        verificationStatus:
          maxLength: 50
          minLength: 0
          type: string
        verifiedBy:
          type: string
          format: uuid
        verifiedAt:
          type: string
          format: date-time
        rejectionReason:
          maxLength: 500
          minLength: 0
          type: string
        expiryDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationResponseApplicationExceptionDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ApplicationExceptionDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
    PaginationResponseApplicationDocumentDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ApplicationDocumentDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...